requirement.txt.txt
Flask
mysql-connector-python
pandas
openpyxl
reportlab